requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
aiosqlite>=0.19.0
yt-dlp>=2024.1.1
openai-whisper>=20231117
anthropic>=0.40.0
python-multipart>=0.0.6
jinja2>=3.1.3
pydantic>=2.5.0
pydantic-settings>=2.1.0
